코드 내용 수정 없이 구조만 바꿈
- 크게 users와 translator로 나눔
- ciceron_lib을 auzre.py로 수정
- bot_daemon/functions.py를 actions로 수정
- run.py 추가
